vakudatuin, loading, delete alert

+++ src/components/Cards/EventCardTable.vue
@@ -110,7 +110,7 @@
 
             <td class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4 ">
               <ButtonPress @click="EditEvent(i.id)">Edit</ButtonPress>
-              <ButtonPress type="submit" @click="DeleteUser(i.id)">
+              <ButtonPress type="submit" class="bg-red-500 hover:bg-orange-600" @click="DeleteUser(i.id)">
                 Delete</ButtonPress>
 
             </td>
@@ -197,17 +197,26 @@ export default {
     },
     // Method to delete user by ID
     async DeleteUser(id) {
-      try {
-        const response = await axios.delete(`/api/v1/admin/events/delete/${id}`);
-        if (response.data.code === 0) {
-          alert('Event deleted successfully!');
-          this.getEventsInfo(); // Re-fetch the users after successful deletion
-        } else {
-          alert('Error deleting user: ' + response.data.msg);
+      // Confirmation prompt
+      const confirmed = window.confirm("Are you sure you want to delete this user?");
+
+      if (confirmed) {
+
+        try {
+          const response = await axios.delete(`/api/v1/admin/events/delete/${id}`);
+          if (response.data.code === 0) {
+            alert('Event deleted successfully!');
+            this.getEventsInfo(); // Re-fetch the users after successful deletion
+          } else {
+            alert('Error deleting user: ' + response.data.msg);
+          }
+        } catch (error) {
+          console.error('Error deleting user:', error);
+          alert('Failed to delete user.');
         }
-      } catch (error) {
-        console.error('Error deleting user:', error);
-        alert('Failed to delete user.');
+      } else {
+        console.log("Deletion canceled.");
+
       }
     },
